Deal with read-only site-packages on homebrew-installed Python on Mac OS X
A homebrew installed Python on Mac OS X has a read-only `/Library/Python/2.7/site-packages` and a read-write `/usr/local/lib/python2.7/site-packages`.  This version warns the user of the failure to install on the former but continues to successfully install Phantom in the latter.

Added `module_name`,  `master_name`, `zip_filename` to increase portability to other modules.

Switched to `print()` as a function.